+++ src/Configuration.cs
@@ -34,6 +34,7 @@ public partial class MainClass : MelonMod
         Mod.AddToList("Default frame color", "#7dc6e3", "You can set the frame color individually by adding a \"color\" field to the picture's JSON config.", new Tags { });
         Mod.AddToList("Default frame padding", 0.01f, "You can set the frame padding individually by adding a \"padding\" field to the picture's JSON config.", new Tags { });
         Mod.AddToList("Default frame thickness", 0.01f, "You can set the frame thickness individually by adding a \"thickness\" field to the picture's JSON config.", new Tags { });
+        Mod.AddToList("Default frame metallicness", 0f, "You can set the frame's metallic property individually by adding a \"metallic\" field to the picture's JSON config.", new Tags { });
         Mod.AddToList("Default picture size", 0.5f, "This is the default size of the pictures when they spawn. It will not change the pictures that are already positioned.", new Tags { });
         Mod.AddToList("Enable transparency", false, 0, "WARNING: this option adds a lag spike on picture creation.\nTo lower the effect, you can enable transparency individually by adding a boolean field \"alpha\" to the picture's JSON config.", new Tags { });
         Mod.AddToList("Show on camera", true, 0, "If disabled, this will hide all pictures from legacy camera, as well as LIV and Rock Cam.\nYou can hide/show any individual picture by clicking the corresponding button while holding it.", new Tags { });
@@ -65,13 +66,14 @@ public partial class MainClass : MelonMod
         defaultColor = Hex2Color((string)Mod.Settings[0].SavedValue);
         defaultPadding = (float)Mod.Settings[1].SavedValue;
         defaultThickness = (float)Mod.Settings[2].SavedValue;
-        defaultSize = (float)Mod.Settings[3].SavedValue;
-        enableAlpha = (bool)Mod.Settings[4].SavedValue;
-        visibility = (bool)Mod.Settings[5].SavedValue;
-        buttonsVisibility = (bool)Mod.Settings[6].SavedValue;
-        spawningFrequency = (float)Mod.Settings[7].SavedValue;
-        gifSpeed = (float)Mod.Settings[8].SavedValue;
-        gifDecodingFrequency = (float)Mod.Settings[9].SavedValue;
+        defaultMetallicness = (float)Mod.Settings[3].SavedValue;
+        defaultSize = (float)Mod.Settings[4].SavedValue;
+        enableAlpha = (bool)Mod.Settings[5].SavedValue;
+        visibility = (bool)Mod.Settings[6].SavedValue;
+        buttonsVisibility = (bool)Mod.Settings[7].SavedValue;
+        spawningFrequency = (float)Mod.Settings[8].SavedValue;
+        gifSpeed = (float)Mod.Settings[9].SavedValue;
+        gifDecodingFrequency = (float)Mod.Settings[10].SavedValue;
     }
 
     /**